Moved PlotCanvas to its own module: plotcanvas.py
